WIP: REPO-4639 Content conversion failed using Tika
In order to fix this issue we need each T-Engine to be be able to find out what it supports from its
engine_config.json. The interface and abstract class used ACS repo to do this, has been moved to this
project so that it can be shared with T-Engines in the alfresco-transform-core. At the same time the
opportunity has been taken to remove the InLineTransform class which is no longer needed by the ACS repo.